refactor(migrate): split Hermes importer modules
